# Conflicts:
#	app/assets/javascripts/components/courses/presentation_course/main_page_show/survey_button.jsx
#	app/mailers/comment_mailer.rb

+++ app/controllers/course_tips_controller.rb
@@ -3,26 +3,36 @@ class CourseTipsController < ApplicationController
   after_action :broadcast_notification, only: [:create]
 
   def index
-    courseTips = CourseTip.where(:course_id => params['course_id']).includes([:user, :course])
+    courseTips = CourseTip.get_course_tips(params['course_id'])
     json_response(courseTips.to_json(include: [:user, :course]))
   end
 
   def create
-    q = CourseTip.create(tip_params)
-    courseTips = CourseTip.where(:id => q.id).includes([:user, :course])
-    @course = Course.find(tip_params[:course_id])
-    #(@course.users.uniq - [current_user]).each do |user|
-    @course.users.uniq.each do |user|
-      Notification.create(recipient: user, actor: current_user, action: "ha inserito un nuovo", notifiable: q)
+    tip = CourseTip.new(tip_params)
+
+    unless tip.save
+      render_json_validation_error tip
+      return
     end
+
+    Notification.send_notifications(params['course_id'], current_user, "ha inserito un nuovo", tip)
+
+    courseTips = CourseTip.where(:id => tip.id).includes([:user, :course])
     json_response(courseTips.to_json(include: [:user, :course]))
   end
 
   def destroy
-    CourseTip.destroy(params[:id])
+    course_tip = CourseTip.find(params[:id])
+
+    if !course_tip.destroy
+      render_json_validation_error course_tip
+      return
+    end
+
     Notification.where(:notifiable_id => params[:id]).where(:notifiable_type => "CourseTip").destroy_all
     Report.where(:reportable_id => params[:id]).where(:reportable_type => "CourseTip").destroy_all
 
+    head :no_content
   end
 
   def update
@@ -32,17 +42,12 @@ class CourseTipsController < ApplicationController
   end
 
   def reportTip
-    reason = params[:reportReason][:reason]
-    tip = CourseTip.find(params[:id])
-    report = Report.where(:reportable_id => params[:id]).where(:reportable_type => "CourseTip").first
 
-    if (report != nil)
-      UserReport.create!(user_id: current_user.id, report_id: report.id, reason: reason)
-    else
-      r = Report.create(action: "È stata segnalato un", reportable: tip)
-      UserReport.create!(user_id: current_user.id, report_id: r.id, reason: reason)
-    end
-
+    Report.send_report(params[:id], current_user.id,
+                       params[:reportReason][:reason],
+                       CourseTip.find(params[:id]),
+                       "CourseTip",
+                       "È stata segnalato un")
     #end
     respond_to do |format|
       format.json { head :ok }
